pyproject.toml
[tool.poetry]
name = "lingvavestra"
version = "0.1.0"
description = ""
authors = ["Mihail Dragan <mike94.0653@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.24.0.post1"
fastapi = {extras = ["all"], version = "^0.104.1"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
asyncpg = "^0.29.0"
alembic = "^1.13.0"
sqlmodel = "^0.0.14"
bcrypt = "4.0.1"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
starlette-admin = "^0.12.2"
sqlalchemy-file = "^0.6.0"
pillow = "^10.2.0"
fastapi-pagination = "^0.12.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 120
combine_as_imports = true


[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
(
  /(
    | venv
    | _build
    | buck-out
    | build
    | migrations
  )/
)
'''
